DQN-Trading/requirements.txt
numpy>=1.24.3
pandas>=2.0.0
torch>=2.0.0
python-dotenv>=1.0.0
alpaca-trade-api>=3.0.0
matplotlib>=3.7.1
seaborn>=0.12.2
tqdm>=4.65.0
scikit-learn>=1.2.2
plotly>=5.14.1
Pillow>=10.0.0
jupyter>=1.0.0
kaleido>=0.2.1
scipy>=1.10.1
six>=1.16.0
